requirements.txt
anyio==4.6.2.post1
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.95.2
h11==0.14.0
idna==3.10
psycopg2-binary==2.9.6
pydantic==1.10.19
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.18
starlette==0.27.0
typing_extensions==4.12.2
uvicorn==0.22.0
python-jose[cryptography]
requests
boto3
passlib
bcrypt
pyjwt